requirements.txt
Flask==2.3.2 
Flask-SQLAlchemy==3.0.5 
Flask-BCrypt==1.0.1 
Flask-Mail==0.10.0 
PyJWT==2.8.0 
Werkzeug==3.0.1
